pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "chorus_upload"
authors = [
    {name = "Will Ashe"},
    {name = "Tony Pan"},
]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
# python 3.12 is needed for pathlib.Path.glob with case_sensitive=False
requires-python = ">=3.13"
# cloudpathlib 0.18 is needed as 0.20 has problems with constructing paths with cloud credentials leading to
# "azure.core.exceptions.ClientAuthenticationError: Server failed to authenticate the request. Make sure the value of Authorization header is formed correctly including the signature."
dependencies = [
    "numpy",
    "pandas",
    "cloudpathlib == 0.23.0",
    "cloudpathlib[azure] == 0.23.0",
    "azure-identity == 1.25.1",
    "tomli",
    "parse",
    "tabulate"
]

[project.optional-dependencies]
s3 = ["cloudpathlib[s3] == 0.23.0",
]
gs = ["cloudpathlib[gs] == 0.23.0",
]

[tool.setuptools]    
pacakges = ["chorus_upload", "tests.create_test_data"]

